+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Disabled vim modeline processing
 - Force vim folds to be always opened to prevent problems
 - Fix vim-easymotion decorators drifting to the end of line ( #60 )
+- Fix incorrect cursor positions after commands/mappings such as `>gv` ( #141 )
 
 ## [0.0.47]